Create __init__.py
Python Package Definition: Placing an __init__.py file in a directory tells Python that the directory should be treated as a package. This allows you to import modules from that directory/package elsewhere in your project using Python's import statement.

Initialization Code: The __init__.py file can contain any Python code that you want to run when the package is imported. This is useful for setting up package-level variables, importing submodules, or running any initialization code needed for the package.

Namespace Management: It helps manage the namespace of the package. For example, you can use it to specify which modules or symbols the package exports as API, by defining __all__ list.