Update feedback_handler.py
In this updated function:

We import the openpyxl library.
We create a new Excel workbook and worksheet using openpyxl.
We add the privacy statement to the worksheet.
We iterate through the feedback_data dictionary and add each key-value pair to a new row in the worksheet.
We save the workbook to a temporary file named feedback.xlsx.
We create the email message and attach the feedback.xlsx file to the email using the msg.attach method.
We send the email using app.extensions['mail'].send(msg).
Finally, we remove the temporary feedback.xlsx file using os.remove.
With these changes, the email sent by the handle_feedback_submission function will include both the feedback data in the body and an attached Excel file with the feedback data in the desired format.

Note: This implementation creates a new Excel file for each feedback submission. If you want to append the feedback data to an existing Excel file instead, you'll need to modify the code accordingly.